requirements.txt
asgiref==3.7.1
Django==4.2.1
django-filter==23.2
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==18.9.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
toposort==1.10
tzdata==2023.3
